requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
flask-login
flask-admin
flask-bcrypt
email-validator
python-dotenv
pymysql==1.1.0
requests
wtforms
gunicorn
pytest
